requirements.txt.txt

---

## 3. `requirements.txt` content:

```txt
pycryptodome
